+++ lib/src/controller/countries_controller.dart
@@ -1,11 +1,108 @@
+/*
+ * Author: Anton Ustinoff <https://github.com/ziqq> | <a.a.ustinoff@gmail.com>
+ * Date: 24 June 2024
+ */
+
 import 'dart:developer' as dev;
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_simple_country_picker/src/controller/countries_provider.dart';
-import 'package:flutter_simple_country_picker/src/controller/countries_state.dart';
 import 'package:flutter_simple_country_picker/src/localization/localization.dart';
 import 'package:flutter_simple_country_picker/src/model/country.dart';
+
+/// {@template countries_state}
+/// Countries state.
+/// {@endtemplate}
+@immutable
+abstract base class CountriesState {
+  /// {@macro countries_state}
+  CountriesState._(this.countries);
+
+  /// Create a [CountriesState] in loading state.
+  ///
+  /// {@macro countries_state}
+  factory CountriesState.loading(List<Country> countries) =
+      CountriesState$Loading;
+
+  /// Create a [CountriesState] in idle state.
+  ///
+  /// {@macro countries_state}
+  factory CountriesState.idle(List<Country> countries) = CountriesState$Idle;
+
+  /// Create a [CountriesState] in error state.
+  ///
+  /// {@macro countries_state}
+  factory CountriesState.error(List<Country> countries) = CountriesState$Error;
+
+  /// Type of state
+  abstract final String type;
+
+  /// Check if is Processing.
+  bool get isLoading => this is CountriesState$Loading;
+
+  /// Check if is Failed.
+  bool get isError => this is CountriesState$Error;
+
+  /// Check if is Idle.
+  bool get isIdle => this is CountriesState$Idle;
+
+  /// List of countries.
+  final List<Country> countries;
+
+  late final Map<String, Country> _table = {
+    for (final country in countries) country.countryCode: country
+  };
+
+  /// Get country by [countryCode]
+  Country? getByCountryCode(String countryCode) => _table[countryCode];
+
+  @override
+  int get hashCode => Object.hash(type, countries);
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      (other is CountriesState &&
+          type == other.type &&
+          listEquals(countries, other.countries));
+
+  @override
+  String toString() => 'CountriesState.$type{countries: $countries}';
+}
+
+/// {@macro countries_state}
+///
+/// Loading state.
+final class CountriesState$Loading extends CountriesState {
+  /// {@macro countries_state}
+  CountriesState$Loading(super.countries) : super._();
+
+  @override
+  String get type => 'loading';
+}
+
+/// {@macro countries_state}
+///
+/// Idle state.
+final class CountriesState$Idle extends CountriesState {
+  /// {@macro countries_state}
+  CountriesState$Idle(super.countries) : super._();
+
+  @override
+  String get type => 'idle';
+}
+
+/// {@macro countries_state}
+///
+/// Error state.
+final class CountriesState$Error extends CountriesState {
+  /// {@macro countries_state}
+  CountriesState$Error(super.countries) : super._();
+
+  @override
+  String get type => 'error';
+}
 
 /// {@template countries_controller}
 /// Countries controller.
@@ -54,18 +151,17 @@ final class CountriesController extends ValueNotifier<CountriesState> {
   /// Search listener.
   void Function()? _listener;
 
-  /// Get the state.
-  CountriesState get state => value;
-
   /// Original countries list.
-  List<Country> _original = [];
-
-  bool _useGroup = true;
+  List<Country> _original = <Country>[];
 
   /// Use group in countries list or not.
   ///
   /// Defalut is `true`.
   bool get useGroup => _useGroup;
+  bool _useGroup = true;
+
+  /// Current state.
+  CountriesState get state => value;
 
   /// Add listener to search controller and provide localization.
   ///
@@ -170,7 +266,8 @@ final class CountriesController extends ValueNotifier<CountriesState> {
 
   @override
   void dispose() {
-    if (_listener != null) search?.removeListener(_listener!);
+    final listener = _listener;
+    if (listener != null) search?.removeListener(listener);
     search?.dispose();
     super.dispose();
   }